Skill familie-felles TokenConfig til egen fil
